Split off `Cardano.Wallet.DB.Store.WalletState.StoreSpec`
