+++ lib/wallet/cardano-wallet.cabal
@@ -934,6 +934,7 @@ test-suite unit
     Cardano.Wallet.DB.Store.UTxOHistory.TxOutCBORSpec
     Cardano.Wallet.DB.Store.Wallets.LayerSpec
     Cardano.Wallet.DB.Store.Wallets.StoreSpec
+    Cardano.Wallet.DB.Store.WalletState.StoreSpec
     Cardano.Wallet.Delegation.ModelSpec
     Cardano.Wallet.DelegationSpec
     Cardano.Wallet.DummyTarget.Primitive.Types
